Added the some Expections to my handlers

+++ HRPlus.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using HRPlus.Application.Contracts.Presistence;
+using HRPlus.Application.Expections;
 using MediatR;
 
 namespace HRPlus.Application.Features.LeaveType.Commands.DeleteLeaveType
@@ -13,6 +14,10 @@ namespace HRPlus.Application.Features.LeaveType.Commands.DeleteLeaveType
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var LeaveTypeToDeleate = await _leaveTypeRepository.GetByIdAsync(request.Id);
+
+            // very the new records 
+            if( LeaveTypeToDeleate == null)
+               throw new NotFoundExpection(nameof(LeaveTypeToDeleate), request.Id);
 
             await _leaveTypeRepository.DeleteAsync(LeaveTypeToDeleate.Id);
